mcp_leagues/chat/requirements_bridge.txt
# OpenRouter Discord Bridge Dependencies
# Add these to your existing requirements.txt or install separately

# Core dependencies
httpx>=0.25.0
python-dotenv>=1.0.0

# Optional: For more robust async operations
aiofiles>=23.0.0

# Optional: For structured logging
structlog>=23.0.0

# Optional: For rate limiting and retries
tenacity>=8.0.0